requirements.txt
argparse==1.4.0
black==21.6b0
boto3==1.17.7
botocore==1.20.7
cachetools==4.2.2
captum==0.4.0
catalogue==2.0.4
future==0.18.2
google-auth==1.26.1
google-auth-oauthlib==0.4.2
joblib==1.0.1
gluonnlp==0.10.0
psutil==5.8.0
efficientnet-pytorch==0.7.1
imageio==2.9.0
imageio-ffmpeg==0.4.4
imgaug==0.4.0
Jinja2==3.0.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
ninja==1.10.2
nltk==3.6.2
numpy==1.19.5
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pytorch-lightning==1.4.0
PyYAML==5.3.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sentencepiece==0.1.91
six==1.15.0
spacy==3.0.6
streamlit==0.88.0
streamlit-drawable-canvas==0.8.0
tabulate==0.8.9
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torchensemble==0.1.6
torchmetrics==0.5.0
torchtext==0.9.0
tornado==6.1
tqdm==4.56.0
transformers==4.5.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.0
yarl==1.6.3
zipp==3.4.1